dumpTheVotes() was not filtering inactive voters (#118)
When votes are dumped in order to launch a tally, sometimes these votes
are requested to be dumped filtering out deactivated voters, i.e. those
voters with is_active=False in the auth_user database User table.

In that case, we were not removing inactive voters, so basically all
votes corresponding to registered votes were always being dumped for
tallying, even if this was not requested action.

This error was most probably introduced when the optimization on vote
dumping was added by filtering using a hand-crafted SQL query and
then filtering with the join command around 14 months ago.